refactor(cli): split art into logo, animation, and command files
- console/logo.ts: logo data + frame generation utilities
- console/animation.ts: interactive and intro animation playback
- cli/commands/art.ts: thin command handler

Follows existing convention of putting display/UI code in console/.
